requirements.txt
Flask==3.1.0
Flask-Session
Flask-Login
Werkzeug
gunicorn
sqlalchemy
sqlalchemy-utils
pymssql
azure-ai-documentintelligence
azure-common
azure-core
azure-identity
azure-keyvault-secrets
azure-mgmt-core
azure-mgmt-search
azure-search-documents
azure-storage-blob
fastapi
aiohttp
